width as isize

+++ examples/minifb.rs
@@ -55,6 +55,6 @@ fn blitter_test(mut fb: &mut Framebuffer, bitmaps: &mut Vec<Bitmap>) {
     // For illustration. It's of course not necessary to copy non-moving or non-changing bitmaps on the framebuffer at each frame.
     bitmaps[1].blit(&mut fb).unwrap();
     bitmaps[2].blit(&mut fb).unwrap();
-    if bitmaps[0].x < WIDTH - 10 { bitmaps[0].x = bitmaps[0].x+3; }
+    if bitmaps[0].x < WIDTH as isize - 10 { bitmaps[0].x = bitmaps[0].x+3; }
     fb.draw_fatpixel(WIDTH/2, HEIGHT/2, 4, 0xffffffff).unwrap();
 }
